hmm i am getting lost in the OOness.
oh where oh where is the right place to put the logic? should there be a single sourceo truth?

some refinement to the adapter engine, feels a lot nicer in the emulator like this and I think I have the final control I want now.

next step is just writing the freakin logic and stop messing with listeners and such