gridview is more or less behaving as i was hoping it would now and seems stable up to about 20x20 columns/rows in my emulator so should preform even better on a decent kindle (my son's device)
I don't think the holder pattern buys us anything as the view should not refresh.

Logic for the games need to be written in such a way that a cards state is contained within itself, gvalue will likely become it's flippedness -- or perhaps something else... I don't know getting sleepy.